animator 저장 불러오기 했는데 static 변수때문에 고생중 맴버로 바꾸고 ifdef했는데 검사 해봐야됨

+++ Moonlighter/Program.cpp
@@ -162,7 +162,7 @@ void Program::MakeGrid()
 	p2DRenderer->SetCamera(true);
 
 	//가로
-	for (int i = 0; i < arrayY.size(); i++)
+	for (int i = 0; i < (int)arrayY.size(); i++)
 	{
 		if (arrayY[i] == 0)
 		{
@@ -173,7 +173,7 @@ void Program::MakeGrid()
 		p2DRenderer->DrawLine(D3DXVECTOR2(arrayX[0] , arrayY[i] ), D3DXVECTOR2(arrayX[arrayX.size() - 1] , arrayY[i] ), nullptr, gridColor);
 	}
 	//세로
-	for (int i = 0; i < arrayX.size(); i++)
+	for (int i = 0; i < (int)arrayX.size(); i++)
 	{
 		if (arrayX[i] == 0)
 		{
